fix: fix cyclic dependence in display caused by handlers placement, moved fetcher related files into separate package and divide entities
