+++ bot.py
@@ -7,10 +7,10 @@ from telegram.ext import (
 
 from config import Config
 from states import States
-from fetcher import Fetcher
+from fetcher.fetcher import Fetcher
 from constants import FETCHER
-from handlers.base_handlers import MiscHandler
-from handlers.entity_handlers import (
+from handlers.entity_handlers import MiscHandler
+from handlers.conversation_handlers import (
     CharactersConversationHandler,
     ComicsConversationHandler,
     EventsConversationHandler,
