+++ fetcher/__init__.py
@@ -1,0 +1,4 @@
+from .fetcher import Fetcher
+
+
+__all__ = ['Fetcher']
