+++ fetcher/entities/__init__.py
@@ -1,0 +1,8 @@
+from .comic import Comic
+from .event import Event
+from .series import Series
+from .general import Creator
+from .character import Character
+
+
+__all__ = ['Creator', 'Character', 'Event', 'Comic', 'Series']
