+++ fetcher/entities/general.py
@@ -1,0 +1,8 @@
+class Creator:
+    def __init__(self, resource_uri, name, role):
+        self.resource_uri = resource_uri
+        self.name = name
+        self.role = role
+
+    def __repr__(self):
+        return f"{self.role}: {self.name}"
