+++ fetcher/entities/series.py
@@ -1,0 +1,29 @@
+class Series:
+    def __init__(
+        self,
+        _id,
+        title,
+        description,
+        img_link,
+        detail,
+        resource_uri,
+        start_year,
+        end_year,
+        next_series,
+        previous_series,
+        creators,
+    ):
+        self._id = _id
+        self.title = title
+        self.description = description
+        self.img_link = img_link
+        self.detail = detail["url"].split("?utm")[0] if detail else ""
+        self.resource_uri = resource_uri
+        self.start_year = start_year
+        self.end_year = end_year
+        self.next_series = next_series
+        self.previous_series = previous_series
+        self.creators = creators
+
+    def __repr__(self):
+        return f"Series(_id: {self._id}, title: {self.title}, resource_uri: {self.resource_uri})"
