+++ handlers/conversation_handlers/__init__.py
@@ -1,0 +1,12 @@
+from .comics import ComicsConversationHandler
+from .events import EventsConversationHandler
+from .series import SeriesConversationHandler
+from .characters import CharactersConversationHandler
+
+
+__all__ = [
+    "CharactersConversationHandler",
+    "ComicsConversationHandler",
+    "EventsConversationHandler",
+    "SeriesConversationHandler",
+]
