+++ handlers/entity_handlers/characters.py
@@ -2,12 +2,10 @@ from telegram import Update
 from telegram.ext import CallbackContext
 
 from text import Text
-from fetcher import Route
+from fetcher.fetcher import Route
 from states import States
-from display import Display
 from custom_keyboard import CustomKeyboard
-from handlers.base_handlers import BaseHandler
-from handlers.conversation_handler_builder import ConversationHandlerBuilder, ConversationHandlerInterface
+from handlers.entity_handlers.base_handler import BaseHandler
 
 
 class CharactersHandler(BaseHandler):
@@ -73,49 +71,4 @@ class CharactersHandler(BaseHandler):
             CustomKeyboard.characters_menu,
             Text.characters_menu,
             States.FIND_CHARACTER_BY_NAME_BEGINNING.value,
-        )
-
-
-class CharactersConversationHandler(ConversationHandlerInterface):
-    @classmethod
-    def _build(cls):
-        menu = {
-            f"^{States.LIST_CHARACTERS.value}$": CharactersHandler.list_,
-            f"^{States.FIND_CHARACTER_BY_NAME.value}$": CharactersHandler.find_by_name,
-            f"^{States.FIND_CHARACTER_BY_NAME_BEGINNING.value}$": CharactersHandler.find_by_name_beginning,
-        }
-        back = {f"^{States.BACK.value}$": CharactersHandler.menu}
-        send = {
-            f"^(?!{States.END.value}).+$": Display.send_character,
-        }
-        list_ = {
-            **back,
-            f"^{States.NEXT_PAGE.value}$": CharactersHandler.list_,
-            f"^{States.PREV_PAGE.value}$": CharactersHandler.list_previous,
-            **send,
-        }
-        find_by_name = {
-            **menu,
-            **back,
-            f"^{States.NEXT_PAGE.value}$": CharactersHandler.find_by_name,
-            f"^{States.PREV_PAGE.value}$": CharactersHandler.list_previous_from_name,
-            **send,
-        }
-        find_by_name_beginning = {
-            **menu,
-            **back,
-            f"^{States.NEXT_PAGE.value}$": CharactersHandler.find_by_name_beginning,
-            f"^{States.PREV_PAGE.value}$": CharactersHandler.list_previous_from_name_beginning,
-            **send,
-        }
-
-        return ConversationHandlerBuilder.build_inner_conversation_handler(
-            {f"^{States.CHARACTERS.value}$": CharactersHandler.menu},
-            {
-                States.CHARACTERS.value: menu,
-                States.LIST_CHARACTERS.value: list_,
-                States.FIND_CHARACTER_BY_NAME.value: find_by_name,
-                States.FIND_CHARACTER_BY_NAME_BEGINNING.value: find_by_name_beginning,
-            },
-            CharactersHandler.save_input,
         )
