+++ handlers/entity_handlers/events.py
@@ -2,12 +2,10 @@ from telegram import Update
 from telegram.ext import CallbackContext
 
 from text import Text
-from fetcher import Route
 from states import States
-from display import Display
+from fetcher.fetcher import Route
 from custom_keyboard import CustomKeyboard
-from handlers.base_handlers import BaseHandler
-from handlers.conversation_handler_builder import ConversationHandlerBuilder, ConversationHandlerInterface
+from handlers.entity_handlers.base_handler import BaseHandler
 
 
 class EventsHandler(BaseHandler):
@@ -71,49 +69,3 @@ class EventsHandler(BaseHandler):
     def list_previous_from_name(cls, update: Update, context: CallbackContext):
         cls._list_previous(context)
         return cls.find_by_name(update, context)
-
-
-class EventsConversationHandler(ConversationHandlerInterface):
-    @classmethod
-    def _build(cls):
-        menu = {
-            f"^{States.LIST_EVENTS.value}$": EventsHandler.list_,
-            f"^{States.FIND_EVENT_BY_NAME.value}$": EventsHandler.find_by_name,
-            f"^{States.FIND_EVENT_BY_NAME_BEGINNING.value}$": EventsHandler.find_by_name_beginning,
-        }
-        back = {
-            f"^{States.BACK.value}$": EventsHandler.menu,
-        }
-        send = {
-            f"^(?!{States.END.value}).+$": Display.send_event,
-        }
-        list_ = {
-            **back,
-            f"^{States.NEXT_PAGE.value}$": EventsHandler.list_,
-            f"^{States.PREV_PAGE.value}$": EventsHandler.list_previous,
-            **send,
-        }
-        find_by_name = {
-            **menu,
-            **back,
-            f"^{States.NEXT_PAGE.value}$": EventsHandler.find_by_name,
-            f"^{States.PREV_PAGE.value}$": EventsHandler.list_previous_from_name,
-            **send,
-        }
-        find_by_name_beginning = {
-            **menu,
-            **back,
-            f"^{States.NEXT_PAGE.value}$": EventsHandler.find_by_name_beginning,
-            f"^{States.PREV_PAGE.value}$": EventsHandler.list_previous_from_name_beginning,
-            **send,
-        }
-        return ConversationHandlerBuilder.build_inner_conversation_handler(
-            {f"^{States.EVENTS.value}$": EventsHandler.menu},
-            {
-                States.EVENTS.value: menu,
-                States.LIST_EVENTS.value: list_,
-                States.FIND_EVENT_BY_NAME.value: find_by_name,
-                States.FIND_EVENT_BY_NAME_BEGINNING.value: find_by_name_beginning,
-            },
-            EventsHandler.save_input,
-        )
